moved the colors to the ppu module
