+++ lib_gbc/src/machine/registers_handler.rs
@@ -5,8 +5,8 @@ use crate::mmu::io_ports::DIV_REGISTER_INDEX;
 use crate::ppu::gb_ppu::GbPpu;
 use crate::utils::bit_masks::*;
 use crate::utils::memory_registers::*;
-use crate::utils::colors::*;
-use crate::utils::color::Color;
+use crate::ppu::colors::*;
+use crate::ppu::color::Color;
 use super::interrupts_handler::InterruptsHandler;
 use crate::keypad::joypad::Joypad;
 use crate::keypad::button::Button;
