+++ lib_gbc/src/ppu/color.rs
