+++ lib_gbc/src/ppu/mod.rs
@@ -1,5 +1,7 @@
 pub mod gb_ppu;
 pub mod ppu_state;
+pub mod color;
+pub mod colors;
 mod normal_sprite;
 mod sprite_attribute;
 mod extended_sprite;
