+++ lib_gbc/src/utils/mod.rs
@@ -1,5 +1,3 @@
 pub mod vec2;
 pub mod memory_registers;
-pub mod color;
-pub mod colors;
 pub mod bit_masks;
